Feat/profile skip

+++ src/main/java/core/domain/bookmark/service/impl/BookmarkServiceImpl.java
@@ -48,6 +48,14 @@ public class BookmarkServiceImpl implements BookmarkService {
     public CursorPageResponse<BookmarkItem> getMyBookmarks( int size, String  cursor) {
         String email = SecurityContextHolder.getContext().getAuthentication().getName();
 
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND));
+
+        if(user.getBirthdate()==null||user.getPurpose()==null||user.getIntroduction()==null||user.getLanguage()==null||user.getHobby()==null||user.getSex()==null){
+            throw new BusinessException(ErrorCode.PROFILE_SET_NOT_COMPLETED);
+        }
+
+
         Pageable pageable = PageRequest.of(0, size + 1);
 
         Long cursorId = null;
@@ -168,13 +176,19 @@ public class BookmarkServiceImpl implements BookmarkService {
     public void addBookmark( Long postId) {
         String email = SecurityContextHolder.getContext().getAuthentication().getName();
 
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND));
+
+        if(user.getBirthdate()==null||user.getPurpose()==null||user.getIntroduction()==null||user.getLanguage()==null||user.getHobby()==null||user.getSex()==null){
+            throw new BusinessException(ErrorCode.PROFILE_SET_NOT_COMPLETED);
+        }
+
+
         Optional<Bookmark> bookmark = bookmarkRepository.findByUserEmailAndPostId(email, postId);
         if (bookmark.isPresent()) {
             throw new BusinessException(ErrorCode.BOOKMARK_ALREADY_EXIST);
         }
 
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND));
         Post post = postRepository.findById(postId).
                 orElseThrow(() -> new BusinessException(ErrorCode.POST_NOT_FOUND));
 
@@ -185,6 +199,14 @@ public class BookmarkServiceImpl implements BookmarkService {
     @Transactional
     public void removeBookmark( Long postId) {
         String email = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND));
+
+        if(user.getBirthdate()==null||user.getPurpose()==null||user.getIntroduction()==null||user.getLanguage()==null||user.getHobby()==null||user.getSex()==null){
+            throw new BusinessException(ErrorCode.PROFILE_SET_NOT_COMPLETED);
+        }
+
 
         bookmarkRepository.deleteByUserEmailAndPostId(email, postId);
     }
